+++ RiskOfChaos/EffectDefinitions/Character/AspectRoulette.cs
@@ -1,6 +1,9 @@
 ﻿using HarmonyLib;
+using RiskOfChaos.ConfigHandling;
 using RiskOfChaos.EffectHandling.EffectClassAttributes;
+using RiskOfChaos.EffectHandling.EffectClassAttributes.Data;
 using RiskOfChaos.Utilities;
+using RiskOfOptions.OptionConfigs;
 using RoR2;
 using System;
 using UnityEngine;
@@ -11,6 +14,13 @@ namespace RiskOfChaos.EffectDefinitions.Character
     [ChaosTimedEffect(90f, AllowDuplicates = false)]
     public sealed class AspectRoulette : TimedEffect
     {
+        [EffectConfig]
+        static readonly ConfigHolder<bool> _allowDirectorUnavailableElites =
+            ConfigFactory<bool>.CreateConfig("Ignore Elite Selection Rules", false)
+                               .Description("If the effect should ignore normal elite selection rules. If enabled, any elite type can be selected, if disabled, only the elite types that can currently be spawned on the stage can be selected")
+                               .OptionConfig(new CheckBoxConfig())
+                               .Build();
+
         [RequireComponent(typeof(CharacterBody))]
         class RandomlySwapAspect : MonoBehaviour
         {
@@ -50,7 +60,7 @@ namespace RiskOfChaos.EffectDefinitions.Character
                 {
                     if (EliteUtils.HasAnyAvailableEliteEquipments)
                     {
-                        inventory.SetEquipmentIndex(EliteUtils.GetRandomEliteEquipmentIndex());
+                        inventory.SetEquipmentIndex(EliteUtils.SelectEliteEquipment(_allowDirectorUnavailableElites.Value));
                     }
                     else
                     {